Make „Featured Sites“ category link scroll to <FilteredShowcase>
Move „Featured Sites“ styles around, too. Missed that in 4582fc4.
