+++ www/src/views/showcase/index.js
@@ -1,26 +1,13 @@
 import React, { Component } from "react"
 import Helmet from "react-helmet"
-import hex2rgba from "hex2rgba"
-import { style } from "glamor"
 
-import FilteredShowcase from "./filtered-showcase"
 import FeaturedSites from "./featured-sites"
+import FilteredShowcase from "./filtered-showcase"
 import Layout from "../../components/layout"
-import { options, rhythm } from "../../utils/typography"
-import presets, { colors } from "../../utils/presets"
-import scrollToAnchor from "../../utils/scroll-to-anchor"
+import presets from "../../utils/presets"
 
 class ShowcaseView extends Component {
   showcase = React.createRef()
-
-  onClickHandler = (target, updateQuery) =>
-    target.current
-      ? scrollToAnchor(target.current, () => {
-          updateQuery(({ filters }) => {
-            return { filters: [`Featured`] }
-          })
-        })
-      : () => {}
 
   render = () => {
     const { location, data } = this.props
@@ -31,9 +18,7 @@ class ShowcaseView extends Component {
           <title>Showcase</title>
         </Helmet>
         <FeaturedSites
-          onClickHandler={this.onClickHandler}
           featured={data.featured.edges}
-          styles={styles}
           showcase={this.showcase}
         />
         <div
@@ -52,86 +37,3 @@ class ShowcaseView extends Component {
 }
 
 export default ShowcaseView
-
-const styles = {
-  featuredSitesCard: style({
-    display: `flex`,
-    flexDirection: `column`,
-    flexGrow: 0,
-    flexShrink: 0,
-    width: 320,
-    marginBottom: rhythm(options.blockMarginBottom * 2),
-    marginRight: rhythm(3 / 4),
-    [presets.Hd]: {
-      width: 360,
-      marginRight: rhythm(6 / 4),
-    },
-    [presets.VHd]: {
-      width: 400,
-    },
-  }),
-  withTitleHover: style({
-    "& .title": {
-      transition: `box-shadow .3s cubic-bezier(.4,0,.2,1), transform .3s cubic-bezier(.4,0,.2,1)`,
-      boxShadow: `inset 0 0px 0px 0px ${colors.ui.whisper}`,
-    },
-    "&:hover .title": {
-      boxShadow: `inset 0 -3px 0px 0px ${colors.ui.bright}`,
-    },
-  }),
-  button: {
-    border: 0,
-    borderRadius: presets.radius,
-    cursor: `pointer`,
-    fontFamily: options.headerFontFamily.join(`,`),
-    fontWeight: `bold`,
-    padding: `${rhythm(1 / 5)} ${rhythm(2 / 3)}`,
-    WebkitFontSmoothing: `antialiased`,
-    "&&": {
-      backgroundColor: colors.gatsby,
-      borderBottom: `none`,
-      boxShadow: `none`,
-      color: `white`,
-      "&:hover": {
-        backgroundColor: colors.gatsby,
-      },
-    },
-  },
-  sticky: {
-    paddingTop: rhythm(options.blockMarginBottom),
-    position: `sticky`,
-    top: 0,
-    [presets.Desktop]: {
-      top: `calc(${presets.headerHeight} - 1px)`,
-    },
-  },
-  scrollbar: {
-    WebkitOverflowScrolling: `touch`,
-    "&::-webkit-scrollbar": {
-      width: `6px`,
-      height: `6px`,
-    },
-    "&::-webkit-scrollbar-thumb": {
-      background: colors.ui.bright,
-    },
-    "&::-webkit-scrollbar-track": {
-      background: colors.ui.light,
-    },
-  },
-  screenshot: {
-    borderRadius: presets.radius,
-    boxShadow: `0 4px 10px ${hex2rgba(colors.gatsby, 0.1)}`,
-    marginBottom: rhythm(options.blockMarginBottom / 2),
-    transition: `all ${presets.animation.speedDefault} ${
-      presets.animation.curveDefault
-    }`,
-  },
-  screenshotHover: {
-    background: `transparent`,
-    color: colors.gatsby,
-    "& .gatsby-image-wrapper": {
-      transform: `translateY(-3px)`,
-      boxShadow: `0 8px 20px ${hex2rgba(colors.lilac, 0.5)}`,
-    },
-  },
-}
